1. add vdd install and uninstall shell bat. 2. optimize script make workspace is current directory.

+++ Community Scripts/silent-install.ps1
@@ -14,31 +14,42 @@ param(
 );
 
 # Create temp directory
-$tempDir = Join-Path $env:TEMP "VDDInstall";
+$tempDir = $PWD;
 New-Item -ItemType Directory -Path $tempDir -Force | Out-Null;
-
-# Download and run DevCon Installer
-Write-Host "Installing DevCon..." -ForegroundColor Cyan;
-$devconPath = Join-Path $tempDir "Devcon.Installer.exe";
-Invoke-WebRequest -Uri "https://github.com/Drawbackz/DevCon-Installer/releases/download/1.4-rc/Devcon.Installer.exe" -OutFile $devconPath;
-Start-Process -FilePath $devconPath -ArgumentList "install -hash $DevconHash -update -dir `"$tempDir`"" -Wait -NoNewWindow;
-
 # Define path to devcon executable
 $devconExe = Join-Path $tempDir "devcon.exe";
+# Download and run DevCon Installer
+if (-not (Test-Path $devconExe))
+{
+    $devconPath = Join-Path $tempDir "Devcon.Installer.exe";
+    if (-not (Test-Path $devconPath))
+    {
+        Write-Host "Downloading DevCon..." -ForegroundColor Cyan;
+        Invoke-WebRequest -Uri "https://github.com/Drawbackz/DevCon-Installer/releases/download/1.4-rc/Devcon.Installer.exe" -OutFile $devconPath;
+    }
+    Write-Host "Installing DevCon..." -ForegroundColor Cyan;
+    Start-Process -FilePath $devconPath -ArgumentList "install -hash $DevconHash -update -dir `"$tempDir`"" -Wait -NoNewWindow;
+    Write-Host "Installing DevCon Completed..." -ForegroundColor Cyan;
+}
 
 # Check if VDD is installed. Or else, install it
 $check = & $devconExe find "Root\MttVDD";
 if ($check -match "1 matching device\(s\) found") {
     Write-Host "Virtual Display Driver already present. No installation." -ForegroundColor Green;
 } else {
-    # Download and unzip VDD
-    Write-Host "Downloading VDD..." -ForegroundColor Cyan;
-    $driverZipPath = Join-Path $tempDir 'driver.zip';
-    Invoke-WebRequest -Uri $DriverURL -OutFile $driverZipPath;
-    Expand-Archive -Path $driverZipPath -DestinationPath $tempDir -Force;
-
     # Extract the signPath certificates
     $catFile = Join-Path $tempDir 'VirtualDisplayDriver\mttvdd.cat';
+    if (-not (Test-Path $catFile)){
+        # Download and unzip VDD
+        $driverZipPath = Join-Path $tempDir 'driver.zip';
+        if (-not (Test-Path $driverZipPath))
+        {
+            Write-Host "Downloading VDD..." -ForegroundColor Cyan;
+            Invoke-WebRequest -Uri $DriverURL -OutFile $driverZipPath;
+        }
+        Expand-Archive -Path $driverZipPath -DestinationPath $tempDir -Force;
+    }
+
     $signature = Get-AuthenticodeSignature -FilePath $catFile;
     $catBytes = [System.IO.File]::ReadAllBytes($catFile);
     $certificates = New-Object System.Security.Cryptography.X509Certificates.X509Certificate2Collection;
@@ -46,8 +57,10 @@ if ($check -match "1 matching device\(s\) found") {
 
     # Create the temp directory for certificates
     $certsFolder = Join-Path $tempDir "ExportedCerts";
-    New-Item -ItemType Directory -Path $certsFolder;
-
+    if (-not (Test-Path $certsFolder))
+    {
+        New-Item -ItemType Directory -Path $certsFolder;
+    }
     # Write and store the driver certificates on local machine
     Write-Host "Installing driver certificates on local machine." -ForegroundColor Cyan;
     foreach ($cert in $certificates) {
